generar agenda plenaria terminada

See merge request JonatanHenriquez/siarcaf!123

+++ app/Http/Controllers/JuntaDirectivaController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers;
 
 
 
+use App\Clases\Mensaje;
 use Carbon\Carbon;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Redirector;
 
@@ -189,41 +191,45 @@ class JuntaDirectivaController extends Controller
     }
 
     public function generar_agenda_plenaria_jd(Request $request, Redirector $redirect)
-    {   //dd($request->all());
-        $agenda =  new Agenda();
-        $agenda->codigo = $request->codigo;
-        $agenda->periodo_id = Periodo::latest()->first()->id;
-        $agenda->lugar = $request->lugar;
-        $agenda->fecha = DateTime::createFromFormat('d/m/Y', $request->fecha)->format('d-m-y'); 
-        $agenda->inicio = DateTime::createFromFormat('d/m/Y H:i:s' , $request->fecha.''.date('H:i:s', strtotime($request->hora)))->format('Y-m-d H:i:s');
-        $agenda->trascendental = 0;
-        if($request->trascendental == 'on')
-            $agenda->trascendental = 1;
+    {
+        if ($request->ajax()){
+            $agenda =  new Agenda();
+            $agenda->codigo = $request->codigo;
+            $agenda->periodo_id = Periodo::latest()->first()->id;
+            $agenda->lugar = $request->lugar;
+            $agenda->fecha = (DateTime::createFromFormat('d-m-Y', $request->fecha))->format('Y-m-d');
+            $agenda->inicio = DateTime::createFromFormat('d-m-Y h:i A' , $request->fecha.' '.date('h:i A', strtotime($request->hora)))->format('Y-m-d H:i:s');
+            $agenda->trascendental = 0;
+            if($request->trascendental == 'on')
+                $agenda->trascendental = 1;
 
-        $agenda->vigente = 1;
-        $agenda->activa = 0;
-        $agenda->fijada = 0;
-        $agenda->save();
+            $agenda->vigente = 1;
+            $agenda->activa = 0;
+            $agenda->fijada = 0;
+            $agenda->save();
 
+            $agendas = Agenda::where('id','!=',0)->orderBy('updated_at','DESC')->get();
+            /*return view('jdagu.listado_agenda_plenaria_jd')
+                ->with('agendas', $agendas);*/
+            $respuesta = new \stdClass();
+            $respuesta->mensaje = (new Mensaje("Exito","Agenda creada con exito","success"))->toArray();
+            return new JsonResponse($respuesta);
 
-        $agendas = Agenda::where('id','!=',0)->orderBy('updated_at','DESC')->get();
-
-        return view('jdagu.listado_agenda_plenaria_jd')
-        ->with('agendas', $agendas);
-
+        }
     }
 
     
     public function eliminar_agenda_creada_jd(Request $request, Redirector $redirect)
     {   //dd($request->all());
 
-        $agenda = Agenda::where('id','=',$request->id_agenda)->first(); 
-        $agenda->delete();
-        $agendas = Agenda::where('id','!=',0)->orderBy('updated_at','DESC')->get();
-
-        return view('jdagu.listado_agenda_plenaria_jd')
-        ->with('agendas', $agendas);
-
+        if ($request->ajax()){
+            $agenda = Agenda::where('id','=',$request->id_agenda)->first();
+            $agenda->delete();
+            $agendas = Agenda::where('id','!=',0)->orderBy('updated_at','DESC')->get();
+            $respuesta = new \stdClass();
+            $respuesta->mensaje = (new Mensaje("Exito","Agenda eliminada con exito","error"))->toArray();
+            return new JsonResponse($respuesta);
+        }
     }
